Move Bootloaders::Bootloader class
