+++ service/lib/agama/storage/manager.rb
@@ -219,7 +219,7 @@ module Agama
 
       # Available bootloaders in the system.
       #
-      # @ return [Array<Bootloaders::Bootloader>]
+      # @ return [Array<Bootloader>]
       def available_bootloaders
         bootloader.available_bootloaders
       end
