+++ service/test/agama/storage/bootloader_prober_test.rb
@@ -21,7 +21,6 @@
 
 require_relative "../../test_helper"
 require "agama/storage/bootloader_prober"
-require "agama/storage/bootloaders"
 
 describe Agama::Storage::BootloaderProber do
   subject { described_class.new }
@@ -50,7 +49,7 @@ describe Agama::Storage::BootloaderProber do
       it "returns only grub2" do
         bootloaders = subject.probe
         expect(bootloaders.size).to eq(1)
-        expect(bootloaders.first).to be_a(Agama::Storage::Bootloaders::Bootloader)
+        expect(bootloaders.first).to be_a(Agama::Storage::Bootloader)
         expect(bootloaders.first.type).to eq(Agama::Storage::BootloaderType::GRUB2)
       end
     end
@@ -67,7 +66,7 @@ describe Agama::Storage::BootloaderProber do
       it "returns only grub2" do
         bootloaders = subject.probe
         expect(bootloaders.size).to eq(1)
-        expect(bootloaders.first).to be_a(Agama::Storage::Bootloaders::Bootloader)
+        expect(bootloaders.first).to be_a(Agama::Storage::Bootloader)
         expect(bootloaders.first.type).to eq(Agama::Storage::BootloaderType::GRUB2)
       end
     end
@@ -84,7 +83,7 @@ describe Agama::Storage::BootloaderProber do
       it "returns only grub2" do
         bootloaders = subject.probe
         expect(bootloaders.size).to eq(1)
-        expect(bootloaders.first).to be_a(Agama::Storage::Bootloaders::Bootloader)
+        expect(bootloaders.first).to be_a(Agama::Storage::Bootloader)
         expect(bootloaders.first.type).to eq(Agama::Storage::BootloaderType::GRUB2)
       end
     end
@@ -104,13 +103,13 @@ describe Agama::Storage::BootloaderProber do
           it "returns grub2, grub2_bls, and systemd_boot with TPM" do
             bootloaders = subject.probe
             expect(bootloaders.size).to eq(3)
-            expect(bootloaders[0]).to be_a(Agama::Storage::Bootloaders::Bootloader)
+            expect(bootloaders[0]).to be_a(Agama::Storage::Bootloader)
             expect(bootloaders[0].type).to eq(Agama::Storage::BootloaderType::GRUB2)
             expect(bootloaders[0].tpm_encryption_auth?).to eq(true)
-            expect(bootloaders[1]).to be_a(Agama::Storage::Bootloaders::Bootloader)
+            expect(bootloaders[1]).to be_a(Agama::Storage::Bootloader)
             expect(bootloaders[1].type).to eq(Agama::Storage::BootloaderType::GRUB2_BLS)
             expect(bootloaders[1].tpm_encryption_auth?).to eq(true)
-            expect(bootloaders[2]).to be_a(Agama::Storage::Bootloaders::Bootloader)
+            expect(bootloaders[2]).to be_a(Agama::Storage::Bootloader)
             expect(bootloaders[2].type).to eq(Agama::Storage::BootloaderType::SYSTEMD_BOOT)
             expect(bootloaders[2].tpm_encryption_auth?).to eq(true)
           end
@@ -125,13 +124,13 @@ describe Agama::Storage::BootloaderProber do
           it "returns grub2, grub2_bls, and systemd_boot without TPM" do
             bootloaders = subject.probe
             expect(bootloaders.size).to eq(3)
-            expect(bootloaders[0]).to be_a(Agama::Storage::Bootloaders::Bootloader)
+            expect(bootloaders[0]).to be_a(Agama::Storage::Bootloader)
             expect(bootloaders[0].type).to eq(Agama::Storage::BootloaderType::GRUB2)
             expect(bootloaders[0].tpm_encryption_auth?).to eq(false)
-            expect(bootloaders[1]).to be_a(Agama::Storage::Bootloaders::Bootloader)
+            expect(bootloaders[1]).to be_a(Agama::Storage::Bootloader)
             expect(bootloaders[1].type).to eq(Agama::Storage::BootloaderType::GRUB2_BLS)
             expect(bootloaders[1].tpm_encryption_auth?).to eq(false)
-            expect(bootloaders[2]).to be_a(Agama::Storage::Bootloaders::Bootloader)
+            expect(bootloaders[2]).to be_a(Agama::Storage::Bootloader)
             expect(bootloaders[2].type).to eq(Agama::Storage::BootloaderType::SYSTEMD_BOOT)
             expect(bootloaders[2].tpm_encryption_auth?).to eq(false)
           end
@@ -146,13 +145,13 @@ describe Agama::Storage::BootloaderProber do
         it "returns grub2, grub2_bls, and systemd_boot without TPM" do
           bootloaders = subject.probe
           expect(bootloaders.size).to eq(3)
-          expect(bootloaders[0]).to be_a(Agama::Storage::Bootloaders::Bootloader)
+          expect(bootloaders[0]).to be_a(Agama::Storage::Bootloader)
           expect(bootloaders[0].type).to eq(Agama::Storage::BootloaderType::GRUB2)
           expect(bootloaders[0].tpm_encryption_auth?).to eq(false)
-          expect(bootloaders[1]).to be_a(Agama::Storage::Bootloaders::Bootloader)
+          expect(bootloaders[1]).to be_a(Agama::Storage::Bootloader)
           expect(bootloaders[1].type).to eq(Agama::Storage::BootloaderType::GRUB2_BLS)
           expect(bootloaders[1].tpm_encryption_auth?).to eq(false)
-          expect(bootloaders[2]).to be_a(Agama::Storage::Bootloaders::Bootloader)
+          expect(bootloaders[2]).to be_a(Agama::Storage::Bootloader)
           expect(bootloaders[2].type).to eq(Agama::Storage::BootloaderType::SYSTEMD_BOOT)
           expect(bootloaders[2].tpm_encryption_auth?).to eq(false)
         end
@@ -193,7 +192,7 @@ describe Agama::Storage::BootloaderProber do
       it "returns only grub2" do
         bootloaders = subject.probe
         expect(bootloaders.size).to eq(1)
-        expect(bootloaders.first).to be_a(Agama::Storage::Bootloaders::Bootloader)
+        expect(bootloaders.first).to be_a(Agama::Storage::Bootloader)
         expect(bootloaders.first.type).to eq(Agama::Storage::BootloaderType::GRUB2)
       end
     end
